requirements/staging.txt
Django==2.2.5
django-cors-headers==3.1.0
django-extra-fields==1.2.4
django-filter==2.2.0
djangorestframework==3.10.3
pytz==2019.2
sqlparse==0.3.0
Unipath==1.1
Pillow==6.1.0
django-cors-headers==3.1.0